LayoutTests/inspector/debugger/function-details-expected.txt

Tests that Debugger.getFunctionDetails command returns correct location. Bug 71808

Debugger was enabled.

Running: testGetFirstLineFunctionDetails
firstLineFunction type = function
Function details: 
lineNumber: 7
columnNumber: undefined
scriptId is valid: true
name: firstLineFunction
displayName: undefined
inferredName: undefined
scopeChain: n/a

Running: testGetNonFirstLineFunctionDetails
notFirstLineFunction type = function
Function details: 
lineNumber: 12
columnNumber: undefined
scriptId is valid: true
name: notFirstLineFunction
displayName: undefined
inferredName: undefined
scopeChain: n/a

Running: testGetDetailsOfFunctionWithInferredName
obj.m type = function
Function details: 
lineNumber: 16
columnNumber: undefined
scriptId is valid: true
name: undefined
displayName: undefined
inferredName: undefined
scopeChain: n/a

Running: testGetDetailsOfFunctionWithDisplayName
functionWithDisplayName type = function
Function details: 
lineNumber: 19
columnNumber: undefined
scriptId is valid: true
name: functionWithDisplayName
displayName: user-friendly name
inferredName: undefined
scopeChain: n/a

Running: testSmallClosure
smallClosure type = function
Function details: 
lineNumber: 22
columnNumber: undefined
scriptId is valid: true
name: undefined
displayName: undefined
inferredName: undefined
scopeChain: n/a

Running: testBigClosure
bigClosure type = function
Function details: 
lineNumber: 35
columnNumber: undefined
scriptId is valid: true
name: undefined
displayName: undefined
inferredName: undefined
scopeChain: n/a
Debugger was disabled.
